move recordsets JSON to external files
