+++ vinyldns/test-fixtures/recordsets/recordset-changes-list-none.json
@@ -1,0 +1,5 @@
+{
+  "zoneId": "123",
+  "maxItems": 100,
+  "recordSetChanges": []
+}
