+++ vinyldns/test-fixtures/recordsets/recordset-update.json
@@ -1,0 +1,45 @@
+{
+  "zone": {
+    "name": "vinyldnstest.sys.vinyldns.net.",
+    "email": "paul_cleary@foo.com",
+    "status": "Active",
+    "created": "2015-10-30T22:47:38Z",
+    "id": "2d9f4ec0-0596-4040-a953-d14e2cca8982",
+    "connection": {
+      "name": "vinyldnstest.sys.vinyldns.net.",
+      "keyName": "cap_all.vinyldns.com",
+      "key": "xxx",
+      "primaryServer": "int-ddns01.resource.vinyldns.net"
+    },
+    "transferConnection": {
+      "name": "vinyldns.",
+      "keyName": "vinyldns.",
+      "key": "OBF:1:ABC+5",
+      "primaryServer": "127.0.0.1"
+    },
+    "adminGroupId": "c314836d-17db-4a57-b849-eb1feffe0ae7",
+    "acl": {
+      "rules": []
+    }
+  },
+  "recordSet": {
+    "id": "e7739220-2a57-45ef-b69f-fdf42a262b87",
+    "zoneId": "2d9f4ec0-0596-4040-a953-d14e2cca8982",
+    "name": "foo.",
+    "type": "A",
+    "status": "Pending",
+    "created": "2015-11-02T16:08:50Z",
+    "ttl": 200,
+    "records": [
+      {
+        "address": "127.0.0.1"
+      }
+    ],
+    "account": "vinyldns"
+  },
+  "userId": "pclear",
+  "changeType": "Create",
+  "status": "Pending",
+  "created": "2015-11-02T16:08:50Z",
+  "id": "b3d4e0a9-a081-4adc-9a95-3ec2e7d26635"
+}
