+++ vinyldns/test-fixtures/recordsets/recordset.json
@@ -1,0 +1,17 @@
+{
+  "recordSet":{
+    "id":"123",
+    "ownerGroupId":"789",
+    "zoneId":"456",
+    "name":"test-01",
+    "type":"A",
+    "status":"Active",
+    "created":"2015-11-02T13:41:54Z",
+    "updated":"2015-11-02T13:41:57Z",
+    "ttl":200,
+    "records":[{
+      "address":"127.0.0.1"
+    }],
+    "account":"vinyldns"
+  }
+}
