+++ vinyldns/test-fixtures/recordsets/recordsets-list-json-2.json
@@ -1,0 +1,17 @@
+{
+  "maxItems": 1,
+  "recordSets": [{
+    "id": "2",
+    "zoneId": "123",
+    "name": "bind9",
+    "type": "A",
+    "status": "Active",
+    "created": "2015-11-02T14:02:08Z",
+    "updated": "2015-11-02T14:02:09Z",
+    "ttl": 300,
+    "records": [{
+      "address": "127.0.0.1"
+    }],
+    "account": "account-test-2"
+  }]
+}
