+++ vinyldns/test-fixtures/recordsets/recordsets-list-none.json
@@ -1,0 +1,4 @@
+{
+  "maxItems": 100,
+  "recordSets": []
+}
